split derby and hsqldb tests
