+++ testing/src/test/java/org/jdbi/v3/testing/junit5/JdbiDerbyExtensionTest.java
@@ -19,6 +19,7 @@ import java.util.UUID;
 import org.jdbi.v3.core.statement.Query;
 import org.jdbi.v3.testing.junit5.internal.TestingInitializers;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.DisabledInNativeImage;
 import org.junit.jupiter.api.condition.EnabledForJreRange;
 import org.junit.jupiter.api.condition.JRE;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -26,32 +27,16 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @EnabledForJreRange(min = JRE.JAVA_17)
-public class JdbiGenericExtensionTest {
+@DisabledInNativeImage // Derby does not support native-image ref: https://issues.apache.org/jira/browse/DERBY-7108
+public class JdbiDerbyExtensionTest {
 
     @RegisterExtension
     public JdbiExtension derbyExtension = new JdbiGenericExtension("jdbc:derby:memory:" + UUID.randomUUID().toString() + ";create=true")
             .withInitializer(TestingInitializers.usersWithData());
 
-    @RegisterExtension
-    public JdbiExtension hsqldbExtension = new JdbiGenericExtension("jdbc:hsqldb:mem:" + UUID.randomUUID())
-            .withCredentials("username", "password")
-            .withInitializer(TestingInitializers.usersWithData());
-
     @Test
     public void testApacheDerby() {
         List<String> userNames = derbyExtension.getJdbi().withHandle(h -> {
-            try (Query query = h.createQuery("SELECT name FROM users ORDER BY id")) {
-                return query.mapTo(String.class).list();
-            }
-        });
-
-        assertThat(userNames).hasSize(2);
-        assertThat(userNames).containsExactly("Alice", "Bob");
-    }
-
-    @Test
-    public void testHsqldb() {
-        List<String> userNames = hsqldbExtension.getJdbi().withHandle(h -> {
             try (Query query = h.createQuery("SELECT name FROM users ORDER BY id")) {
                 return query.mapTo(String.class).list();
             }
